third_party/blink/web_tests/fast/parser/scriptexec-during-parserInsertBefore-expected.txt

Ensure that DOM is consistent after a specific child has been removed during reparenting.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS containerNode.firstChild is firstChild
PASS nextChild.previousSibling is null
PASS successfullyParsed is true

TEST COMPLETE
